backend/requirements.txt
django
channels
django-redis